Return python set for set types
For set-type columns, the Datum.to_python() and to_json() both
sort their results. It's especially bad with to_python() as it is
called every time the column is accessed and objects like
Port_Group.ports can have thousands of entries. Add to that that
Python has a set type with O(1) lookup, and it doesn't make much
sense for it to be sorting and returning lists with every attribute
access.

This is a backward-incompatible change. Any code that uses
isinstance(row.column, list) or accesses row.column[n] or uses
list operations like append() will break.